Revert "Clear history command"
