+++ package.json
@@ -19,6 +19,6 @@
     "@eslint/js": "^9.20.0",
     "eslint": "^9.25.0",
     "globals": "^15.15.0",
-    "nodemon": "^3.1.10"
+    "nodemon": "^3.1.9"
   }
 }
