--- src/utils/splitMessages.js
@@ -1,10 +1,0 @@
-// Split large AI responses into chunks of messages, so that it can be sent in discord.
-function splitMessage(text, maxLength = 2000) {
-  const chunks = [];
-  for (let i = 0; i < text.length; i += maxLength) {
-    chunks.push(text.slice(i, i + maxLength));
-  }
-  return chunks;
-}
-
-module.exports = splitMessage;
